requirements.txt
-e git+https://github.com/aclarknet/django-xhtml2pdf.git@ca6fae09f70856d701124583cacb9e5a83e7ac26#egg=django_xhtml2pdf
-e git+https://github.com/alex/django-taggit.git@95776ac66948ed7ba7c12e35c1170551e3be66a5#egg=django_taggit
Babel==2.5.1
Django==2.0
Faker==0.8.8
Pillow==4.3.0
PyPDF2==1.26.0
PyYAML==3.12
appdirs==1.4.3
boto==2.48.0
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
diff-match-patch==20121119
dj-database-url==0.4.2
django-admin-sortable2==0.6.18
django-crispy-forms==1.7.0
django-import-export==0.6.1
django-multiselectfield==0.1.8
django-phonenumber-field==1.3.0
django-recaptcha==1.3.1
django-ses==0.8.4
django-solo==1.1.2
djangorestframework==3.7.7
et-xmlfile==1.0.1
geoip2==2.6.0
gunicorn==19.7.1
html5lib==1.0b8
httplib2==0.10.3
idna==2.6
jdcal==1.3
lxml==4.1.1
matplotlib==2.1.1
maxminddb==1.3.0
numpy==1.13.3
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
packaging==16.8
phonenumbers==8.8.8
phonenumberslite==8.8.8
psycopg2==2.7.3.2
pyparsing==2.2.0
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.3
reportlab==3.4.0
requests==2.18.4
six==1.11.0
tablib==0.12.1
text-unidecode==1.1
tmp==0.0.2
unicodecsv==0.14.1
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
xhtml2pdf==0.1b3
xlrd==1.1.0
xlwt==1.3.0
